Add telescreen announcement when the thunderdome is active (#42199)
cl
add: Entertainment telescreens in the bar indicate when they will show something interesting.
tweak: Entertainment telescreens can be viewed from a distance.
/cl

Bargoers are now informed both visually and in chat when clicking on the thunderdome telescreens is likely to actually show something interesting. Additionally, thunderdome telescreens can be watched without being directly adjacent (think sports bar TVs).

+++ code/modules/admin/secrets.dm
@@ -131,7 +131,7 @@
 				for(var/mob/living/mob in thunderdome)
 					qdel(mob) //Clear mobs
 			for(var/obj/obj in thunderdome)
-				if(!istype(obj, /obj/machinery/camera))
+				if(!istype(obj, /obj/machinery/camera) && !istype(obj, /obj/effect/abstract/proximity_checker))
 					qdel(obj) //Clear objects
 
 			var/area/template = GLOB.areas_by_type[/area/tdome/arena_source]
